"Frameworks: Change navbar general behaviour (1/2)
This commit changes that navbar is present on all devices even
with hardware keys. On hardware key devices the navbar height is
set to 0 as default now. Several reasons for that.
- no SystemUI restart when de/activating (last one we had :p)
- our navbar is very resource friendly so even no problem for the
low end devices
- we can now use some of the navbar stuff you will see in next commits
navbar magic :p
This commit also respects the different handling off menu's. Thats
why a part of the code (hasNavigationBar()) will be still left because
it is needed in main for
core/java/android/view/ViewConfiguration.java
core/java/android/view/IWindowManager.aidl
core/java/android/view/WindowManagerPolicy.java
this files which are controling that

Change-Id: I7ca7516d612ff61bf8d5a6ad2acf28202dc2b2a4"